refactor(site/src): split agent settings pages
