+++ site/src/pages/AgentsPage/AgentSettingsExperimentsPage.tsx
@@ -1,0 +1,34 @@
+import type { FC } from "react";
+import { useMutation, useQuery, useQueryClient } from "react-query";
+import {
+	chatDesktopEnabled,
+	updateChatDesktopEnabled,
+} from "#/api/queries/chats";
+import { useAuthenticated } from "#/hooks/useAuthenticated";
+import { RequirePermission } from "#/modules/permissions/RequirePermission";
+import { AgentSettingsExperimentsPageView } from "./AgentSettingsExperimentsPageView";
+
+const AgentSettingsExperimentsPage: FC = () => {
+	const { permissions } = useAuthenticated();
+	const queryClient = useQueryClient();
+	const desktopEnabledQuery = useQuery({
+		...chatDesktopEnabled(),
+		enabled: permissions.editDeploymentConfig,
+	});
+	const saveDesktopEnabledMutation = useMutation(
+		updateChatDesktopEnabled(queryClient),
+	);
+
+	return (
+		<RequirePermission isFeatureVisible={permissions.editDeploymentConfig}>
+			<AgentSettingsExperimentsPageView
+				desktopEnabledData={desktopEnabledQuery.data}
+				onSaveDesktopEnabled={saveDesktopEnabledMutation.mutate}
+				isSavingDesktopEnabled={saveDesktopEnabledMutation.isPending}
+				isSaveDesktopEnabledError={saveDesktopEnabledMutation.isError}
+			/>
+		</RequirePermission>
+	);
+};
+
+export default AgentSettingsExperimentsPage;
