+++ site/src/pages/AgentsPage/AgentsPageView.stories.tsx
@@ -16,7 +16,6 @@ import { API } from "#/api/api";
 import type * as TypesGen from "#/api/typesGenerated";
 import type { Chat } from "#/api/typesGenerated";
 import { DeleteDialog } from "#/components/Dialogs/DeleteDialog/DeleteDialog";
-import { useAuthenticated } from "#/hooks/useAuthenticated";
 import {
 	MockNoPermissions,
 	MockPermissions,
@@ -29,9 +28,13 @@ import {
 import AgentAnalyticsPage from "./AgentAnalyticsPage";
 import AgentCreatePage from "./AgentCreatePage";
 import { AgentSettingsAgentsPageView } from "./AgentSettingsAgentsPageView";
-import { AgentSettingsBehaviorPageView } from "./AgentSettingsBehaviorPageView";
+import AgentSettingsCompactionPage from "./AgentSettingsCompactionPage";
+import AgentSettingsExperimentsPage from "./AgentSettingsExperimentsPage";
+import AgentSettingsGeneralPage from "./AgentSettingsGeneralPage";
+import AgentSettingsLifecyclePage from "./AgentSettingsLifecyclePage";
 import AgentSettingsPage from "./AgentSettingsPage";
 import AgentSettingsSpendPage from "./AgentSettingsSpendPage";
+import AgentSettingsSystemInstructionsPage from "./AgentSettingsSystemInstructionsPage";
 import { AgentsPageView } from "./AgentsPageView";
 import type { ModelSelectorOption } from "./components/ChatElements";
 
@@ -151,59 +154,6 @@ const buildChat = (overrides: Partial<Chat> = {}): Chat => ({
 // across timezones.
 const fixedNow = dayjs("2026-03-12T12:00:00");
 
-// Renders the real PageView components with mock data so the
-// visual snapshots match the actual UI.
-const BehaviorRouteElement = () => {
-	const { permissions } = useAuthenticated();
-	return (
-		<AgentSettingsBehaviorPageView
-			canSetSystemPrompt={permissions.editDeploymentConfig}
-			systemPromptData={{
-				system_prompt: "",
-				include_default_system_prompt: true,
-				default_system_prompt: "You are Coder, an AI coding assistant...",
-			}}
-			planModeInstructionsData={{
-				plan_mode_instructions: "",
-			}}
-			userPromptData={{ custom_prompt: "" }}
-			desktopEnabledData={{ enable_desktop: false }}
-			workspaceTTLData={{ workspace_ttl_ms: 0 }}
-			isWorkspaceTTLLoading={false}
-			isWorkspaceTTLLoadError={false}
-			modelConfigsData={[]}
-			modelConfigsError={undefined}
-			isLoadingModelConfigs={false}
-			thresholds={[]}
-			isThresholdsLoading={false}
-			thresholdsError={undefined}
-			onSaveSystemPrompt={fn()}
-			isSavingSystemPrompt={false}
-			isSaveSystemPromptError={false}
-			onSavePlanModeInstructions={fn()}
-			isSavingPlanModeInstructions={false}
-			isSavePlanModeInstructionsError={false}
-			onSaveUserPrompt={fn()}
-			isSavingUserPrompt={false}
-			isSaveUserPromptError={false}
-			onSaveDesktopEnabled={fn()}
-			isSavingDesktopEnabled={false}
-			isSaveDesktopEnabledError={false}
-			onSaveWorkspaceTTL={fn()}
-			isSavingWorkspaceTTL={false}
-			isSaveWorkspaceTTLError={false}
-			retentionDaysData={{ retention_days: 30 }}
-			isRetentionDaysLoading={false}
-			isRetentionDaysLoadError={false}
-			onSaveRetentionDays={fn()}
-			isSavingRetentionDays={false}
-			isSaveRetentionDaysError={false}
-			onSaveThreshold={fn(async () => undefined)}
-			onResetThreshold={fn(async () => undefined)}
-		/>
-	);
-};
-
 const AgentsRouteElement = () => (
 	<AgentSettingsAgentsPageView
 		exploreModelOverrideData={{ has_malformed_override: false }}
@@ -224,8 +174,19 @@ const agentsRouting = {
 			path: "settings",
 			element: <AgentSettingsPage />,
 			children: [
-				{ index: true, element: <Navigate to="behavior" replace /> },
-				{ path: "behavior", element: <BehaviorRouteElement /> },
+				{ index: true, element: <AgentSettingsGeneralPage /> },
+				{ path: "general", element: <AgentSettingsGeneralPage /> },
+				{
+					path: "behavior",
+					element: <Navigate to="/agents/settings/general" replace />,
+				},
+				{ path: "compaction", element: <AgentSettingsCompactionPage /> },
+				{
+					path: "system-instructions",
+					element: <AgentSettingsSystemInstructionsPage />,
+				},
+				{ path: "experiments", element: <AgentSettingsExperimentsPage /> },
+				{ path: "lifecycle", element: <AgentSettingsLifecyclePage /> },
 				{ path: "agents", element: <AgentsRouteElement /> },
 				{ path: "spend", element: <AgentSettingsSpendPage now={fixedNow} /> },
 				{
@@ -349,10 +310,39 @@ const meta: Meta<typeof AgentsPageView> = {
 		spyOn(API.experimental, "getChatDesktopEnabled").mockResolvedValue({
 			enable_desktop: false,
 		});
+		spyOn(API.experimental, "updateChatDesktopEnabled").mockResolvedValue();
+		spyOn(API.experimental, "getChatPlanModeInstructions").mockResolvedValue({
+			plan_mode_instructions: "",
+		});
+		spyOn(
+			API.experimental,
+			"updateChatPlanModeInstructions",
+		).mockResolvedValue();
+		spyOn(
+			API.experimental,
+			"getUserChatCompactionThresholds",
+		).mockResolvedValue({
+			thresholds: [],
+		});
+		spyOn(
+			API.experimental,
+			"updateUserChatCompactionThreshold",
+		).mockResolvedValue({
+			model_config_id: defaultModelConfigID,
+			threshold_percent: 70,
+		});
+		spyOn(
+			API.experimental,
+			"deleteUserChatCompactionThreshold",
+		).mockResolvedValue();
 		spyOn(API.experimental, "getChatWorkspaceTTL").mockResolvedValue({
 			workspace_ttl_ms: 0,
 		});
 		spyOn(API.experimental, "updateChatWorkspaceTTL").mockResolvedValue();
+		spyOn(API.experimental, "getChatRetentionDays").mockResolvedValue({
+			retention_days: 30,
+		});
+		spyOn(API.experimental, "updateChatRetentionDays").mockResolvedValue();
 		spyOn(API.experimental, "getChatUsageLimitConfig").mockResolvedValue({
 			spend_limit_micros: null,
 			period: "month",
@@ -669,9 +659,7 @@ export const OpensSettingsForAdmins: Story = {
 
 		await waitFor(() => {
 			expect(
-				screen.getByText(
-					"Custom instructions that shape how the agent responds in your conversations.",
-				),
+				screen.getByText("Personal preferences for your chat experience."),
 			).toBeInTheDocument();
 		});
 	},
@@ -689,9 +677,7 @@ export const OpensSettingsForNonAdmins: Story = {
 
 		await waitFor(() => {
 			expect(
-				screen.getByText(
-					"Custom instructions that shape how the agent responds in your conversations.",
-				),
+				screen.getByText("Personal preferences for your chat experience."),
 			).toBeInTheDocument();
 		});
 	},
@@ -707,9 +693,7 @@ export const SettingsViewResets: Story = {
 
 		await waitFor(() => {
 			expect(
-				screen.getByText(
-					"Custom instructions that shape how the agent responds in your conversations.",
-				),
+				screen.getByText("Personal preferences for your chat experience."),
 			).toBeInTheDocument();
 		});
 
@@ -727,13 +711,11 @@ export const SettingsViewResets: Story = {
 		const backButton = screen.getByLabelText("Back to Agents");
 		await userEvent.click(backButton);
 
-		// Re-open settings, should reset to Behavior
+		// Re-open settings, should reset to General
 		await openSettingsView(canvasElement);
 		await waitFor(() => {
 			expect(
-				screen.getByText(
-					"Custom instructions that shape how the agent responds in your conversations.",
-				),
+				screen.getByText("Personal preferences for your chat experience."),
 			).toBeInTheDocument();
 		});
 	},
